+++ CHANGELOG.md
@@ -7,6 +7,13 @@ availableAt:
 ---
 
 ### unreleased
+#### Noelware.Violet
+- Add **violet/Print.h** header to bridge the `<print>` header in C++23 that works in C++20 ([`@auguwu`])
+- Moved all assertion logic into `violet/Language/Assert.h` so that it doesn't have to depend on `violet/Violet.h` if it is not needed ([`@auguwu`])
+
+#### Noelware.Violet.Experimental
+- **SmolString** now uses Violet's assertion logic rather than C's `assert` macro ([`@auguwu`])
+
 **Git History**: <https://github.com/Noelware/violet/compare/26.04.03...master>
 
 ### 26.04.03